Change to location and call to `impute_missing_sigma_bs` function
