+++ echopop/inversion/__init__.py
@@ -1,5 +1,9 @@
 from .inversion_base import InversionBase
 from .inversion_length_TS import InversionLengthTS
+from .operations import impute_missing_sigma_bs
 # from .inversion_matrix_krill import InversionMatrixKrill
 
-__all__ = ["InversionBase", "InversionLengthTS"]
+__all__ = [
+    "InversionBase", "InversionLengthTS",
+    "impute_missing_sigma_bs",
+]
